Fixed bugs and corrected screen size settings
There was a bug in the random generation of buttons in the second and third part of the sound game, because for every animal it generated always the same buttons. Now the issue is fixed: for every different animal the player interacts with always a new sequence of buttons is randomly generated.